+++ src/include/index/flat_l2_index.h
@@ -105,7 +105,7 @@ class flat_l2_index {
     // @todo
   }
 
-  uint32_t dimensions() const {
+  uint64_t dimensions() const {
     return _cpo::dimensions(*feature_vectors_);
   }
